Create letters-in-words.rb
Each other_str parameter defines a set of characters to count. The intersection 
of these sets defines the characters to count in str. Any other_str that starts 
with a caret ^ is negated. The sequence c1-c2 means all characters between c1 and c2. 
The backslash character \ can be used to escape ^ or - and is otherwise ignored 
unless it appears at the end of a sequence or the end of a other_str.

#Write a program that counts the number of a particular letter in the given word. 
#Part 2: Exclude the vowels